README.md
in this study, 7544
news data in the health category of the New York Times were studied in order to
understand the impact of these news on people. The data was pulled using PHP by
collecting the links with the help of the sitemap. News data on natural language processing
(NLP) is applied and sentiment analysis is performed. A project is made with the Python
program for emotion labeling using logistic regression and random forest classification
algorithms and these two algorithms are compared. In the project, there is detailed
information about data analysis, data manipulations, visualizations, statistical approaches,
machine learning and their codes.
As a result, in order to realize the emotional differences between the periods in which the
number of cases showed ups and downs, the sentiments are labeled, classified with
different models and the results were evaluated.